Add human readable IAQ scores
From table 4, page 9 of the BME680 datasheet, I've added in the AQ classification (good, average, etc), tidied up the formatting of the output and made a couple of extremely trivial pylint fixups. My coding skills are still pretty basic and there is probably a much more efficient way to do this, so feel free to reject this if it's unsuitable.