Support for low/high gas variants
The BME688 is a "high" gas resistance variant of the BME680 and requires a slightly different init and gas compensation formula.

This change attempts to auto-detect the chip variant ID and do the right thing.

The constant ENABLE_GAS_MEAS is now set to -1 for back-compat, and passing this into set_gas_status will use the detected variant.